Starting to add the concept of warnings/errors that can be consolidated into a report for the user at the end.
Building an ErrorClass. It is designed to just be a structured object to contain the relevant information about the error that occurred during processing.
This class should be used for cases where a fix can not be automatically applied and requires user input.

Also creating the exception to go along with this ErrorClass. A UserInputRequiredException is thrown at the end for each ErrorClass.

The AliasDict now stores "warnings" and "errors". The idea being that errors are things that we are sure cannot be fixed automatically. And warnings are things that we think we did right, but you should check.