The first usage of the ErrorClass is created.
I found a case where if a QVariant [class object] is used, there is no replacement that could always make sense in apiV2.0.

The case I found was as follows, but I tried to be careful to not scope specifically to that use case in case there are others usages of QVarant[class object] that I could not think of.

>>> if isinstance(myvar, QVariant):
>>>     myvar = myvar.toInt()

If you replace the QVariant with a "type(None)", there is a, small but still important, case where if "myvar" was a reference to None, it would become True, and None does not have a ".toInt()".
The only abstract replacement that I could think of was something like as follows:

>>> if isinstance(myvar, type("QVariant_api1.0_klass", (object,), {})):
>>>     myvar = myvar.toInt()

Which, while technically always correct for the isinstance case, makes a fairly large assumption about all cases where you'd use a QVariant [class object].

I didn't feel comfortable making that assumption with only one data point, so for the time being, we are stating that this statement is un-translateable to api2.0.

We create a ErrorClass with a helpful error message, passing in our node.

We have also fixed the _qstring, _qstringlist, _qchar, _qstringref and _qvariant "psep_process" expressions.