The fix for #19
I found the regression that was introduced in 0.2.5.
It was because I was replacing the entire statement, which in some cases included child nodes that I would want to replace later.
This became obvious in the following case:
>>> size = QtGui.QSizePolicy(QtGui.QSizePolicy.Horrizontal, QtGui.QSizePolicy.Vertical)
In the above statement, we would detect 3 nodes that need to be replaced but as soon as we replaced the outer one, the two child nodes would become orphaned and became null replacements.
The new method that we are using is to replace the least destructive node. In the above case, that would be just the second-level-module, QtGui, for each node which is the first part of it.
That way we are essentially replacing only the child nodes and never the actual root node that may contain more child nodes that we want to replace.

We have overridden the message in _change_verbose to bold our actual replacement part in the entire statement.